renamed components folder to entities folder
