+++ src/app/Game.js
@@ -1,8 +1,8 @@
 import { Application, Assets, Sprite } from 'pixi.js'
 import { CONFIG } from './config.js'
 import { assetsLoader } from '../assets/assetsLoader.js'
-import { Ship } from '../components/Ship.js'
-import { Asteroid } from '../components/Asteroid'
+import { Ship } from '../entities/Ship.js'
+import { Asteroid } from '../entities/Asteroid'
 import { BulletsCounter } from '../ui/bulletsCounter/BulletsCounter'
 import { CountDownTimer } from '../ui/countDownTimer/CountDownTimer'
 import { Button } from '../ui/button/Button'
