+++ src/entities/Asteroid.js
