+++ src/entities/Bullet.js
